REFACTOR into InstPanel, add ChordSpanFill for arps
