+++ midimasterpiece/src/main/java/org/vibehistorian/midimasterpiece/midigenerator/DrumPanel.java
@@ -2,14 +2,11 @@ package org.vibehistorian.midimasterpiece.midigenerator;
 
 import java.awt.event.ActionListener;
 
-import javax.swing.JButton;
 import javax.swing.JCheckBox;
-import javax.swing.JComboBox;
 import javax.swing.JLabel;
-import javax.swing.JPanel;
 import javax.swing.JTextField;
 
-public class DrumPanel extends JPanel {
+public class DrumPanel extends InstPanel {
 	/**
 	 * 
 	 */
@@ -27,26 +24,12 @@ public class DrumPanel extends JPanel {
 	}
 	
 	private JTextField pitch = new JTextField("36", 2);
-	private JTextField hitsPerPattern = new JTextField("8", 2);
-	private JTextField chordSpan = new JTextField("1", 1);
-	
-	private JTextField pauseChance = new JTextField("70", 2);
-	private JTextField exceptionChance = new JTextField("5", 2);
 	
 	private JTextField velocityMin = new JTextField("60", 3);
 	private JTextField velocityMax = new JTextField("100", 3);
 	
-	private JTextField slideMiliseconds = new JTextField("0", 4);
 	private JTextField swingPercent = new JTextField("50", 2);
-	
-	private JTextField patternSeed = new JTextField("0", 8);
-	private JComboBox<String> pattern = new JComboBox<String>();
 	private JCheckBox isVelocityPattern = new JCheckBox("Dynamic", true);
-	private JTextField patternShift = new JTextField("0", 1);
-	
-	private JCheckBox muteInst = new JCheckBox("Mute", false);
-	
-	private JButton removeButton = new JButton("X");
 	
 	public void initComponents() {
 		this.add(new JLabel("#"));
@@ -71,8 +54,8 @@ public class DrumPanel extends JPanel {
 		
 		this.add(new JLabel("Swing(%)"));
 		this.add(swingPercent);
-		this.add(new JLabel("Slide(ms)"));
-		this.add(slideMiliseconds);
+		this.add(new JLabel("Delay(ms)"));
+		this.add(delay);
 		
 		this.add(new JLabel("Seed"));
 		this.add(patternSeed);
@@ -103,38 +86,6 @@ public class DrumPanel extends JPanel {
 		this.pitch.setText(String.valueOf(pitch));
 	}
 	
-	public int getHitsPerPattern() {
-		return Integer.valueOf(hitsPerPattern.getText());
-	}
-	
-	public void setHitsPerPattern(int hitsPerPattern) {
-		this.hitsPerPattern.setText(String.valueOf(hitsPerPattern));
-	}
-	
-	public int getChordSpan() {
-		return Integer.valueOf(chordSpan.getText());
-	}
-	
-	public void setChordSpan(int chordSpan) {
-		this.chordSpan.setText(String.valueOf(chordSpan));
-	}
-	
-	public int getPauseChance() {
-		return Integer.valueOf(pauseChance.getText());
-	}
-	
-	public void setPauseChance(int pauseChance) {
-		this.pauseChance.setText(String.valueOf(pauseChance));
-	}
-	
-	public int getExceptionChance() {
-		return Integer.valueOf(exceptionChance.getText());
-	}
-	
-	public void setExceptionChance(int exceptionChance) {
-		this.exceptionChance.setText(String.valueOf(exceptionChance));
-	}
-	
 	public int getVelocityMin() {
 		return Integer.valueOf(velocityMin.getText());
 	}
@@ -151,34 +102,10 @@ public class DrumPanel extends JPanel {
 		this.velocityMax.setText(String.valueOf(velocityMax));
 	}
 	
-	public int getSlideMiliseconds() {
-		return Integer.valueOf(slideMiliseconds.getText());
-	}
-	
-	public void setSlideMiliseconds(int slideMiliseconds) {
-		this.slideMiliseconds.setText(String.valueOf(slideMiliseconds));
-	}
-	
-	public int getPatternSeed() {
-		return Integer.valueOf(patternSeed.getText());
-	}
-	
-	public void setPatternSeed(int patternSeed) {
-		this.patternSeed.setText(String.valueOf(patternSeed));
-	}
-	
-	public RhythmPattern getPattern() {
-		return RhythmPattern.valueOf((String) pattern.getSelectedItem());
-	}
-	
-	public void setPattern(RhythmPattern pattern) {
-		this.pattern.setSelectedItem((String.valueOf(pattern.toString())));
-	}
-	
 	public DrumPart toDrumPart(int lastRandomSeed) {
 		DrumPart part = new DrumPart(getPitch(), getHitsPerPattern(), getChordSpan(),
 				getPauseChance(), getExceptionChance(), getVelocityMin(), getVelocityMax(),
-				getSlideMiliseconds(), (getPatternSeed() != 0) ? getPatternSeed() : lastRandomSeed,
+				getDelay(), (getPatternSeed() != 0) ? getPatternSeed() : lastRandomSeed,
 				getPattern(), getIsVelocityPattern(), getPatternShift(), getMuteInst(),
 				getSwingPercent());
 		part.setOrder(getPanelOrder());
@@ -197,7 +124,7 @@ public class DrumPanel extends JPanel {
 		setVelocityMin(part.getVelocityMin());
 		setVelocityMax(part.getVelocityMax());
 		
-		setSlideMiliseconds(part.getSlideMiliseconds());
+		setDelay(part.getSlideMiliseconds());
 		setSwingPercent(part.getSwingPercent());
 		
 		setPatternSeed(part.getPatternSeed());
@@ -217,22 +144,6 @@ public class DrumPanel extends JPanel {
 	
 	public void setIsVelocityPattern(boolean isVelocityPattern) {
 		this.isVelocityPattern.setSelected(isVelocityPattern);
-	}
-	
-	public int getPatternShift() {
-		return Integer.valueOf(patternShift.getText());
-	}
-	
-	public void setPatternShift(int shift) {
-		patternShift.setText(String.valueOf(shift));
-	}
-	
-	public boolean getMuteInst() {
-		return muteInst.isSelected();
-	}
-	
-	public void setMuteInst(boolean selected) {
-		this.muteInst.setSelected(selected);
 	}
 	
 	public int getSwingPercent() {
