+++ midimasterpiece/src/main/java/org/vibehistorian/midimasterpiece/midigenerator/MelodyGenerator.java
@@ -671,6 +671,14 @@ public class MelodyGenerator implements JMC {
 							.get(p) == 0) {
 						pitch = Integer.MIN_VALUE;
 					}
+					if (ap.getCspanFill() != ChordSpanFill.ALL) {
+						if ((ap.getCspanFill() == ChordSpanFill.EVEN) && (j % 2 != 0)) {
+							pitch = Integer.MIN_VALUE;
+						}
+						if ((ap.getCspanFill() == ChordSpanFill.ODD) && (j % 2 == 0)) {
+							pitch = Integer.MIN_VALUE;
+						}
+					}
 					
 					arpCPhrase.addChord(new int[] { pitch }, chordDurationArp);
 				}
